requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
pymongo==4.13.2
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.11.7
httpx==0.27.0

# Document processing
PyPDF2==3.0.1
pandas==2.2.2
python-docx==1.1.2
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.2
